Flask/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3